Implemented win/lose conditions
- Pleasure empty = lose the level
- Pleasure full = win the level
- ScoreScreen shows whether you had won or lost the level. If you won,
clicking next will go to the pressure phase. If you lost, clicking next
will go to the main menu.
- Commented the line out that executes when a ScreenSwitchButton
switches to the ScoreScreen, as the ScoreScreen constructor now needs a
parameter and that can't be given when a ScreenSwitchButton redirects
there as of right now. Also, there shouldn't be a button to go to the
score anyways.
- PopupWindow now fires an event "CLOSE_CLICKED" when the close button
is clicked. This is needed to correctly continue after the ActionScreen
paused itself to show the window.